+++ src/server/replicated_archetypes.rs
@@ -67,10 +67,12 @@ impl ReplicatedArchetypes {
                 }
             }
 
-            // Incoming receive markers need to be processed before other components so they can
-            // affect which receive functions are selected for the rest of the entity update.
+            // Incoming same-update markers need to be processed before other components so they
+            // can affect which receive functions are selected for the rest of the entity update.
             replicated_archetype.components.sort_by_key(|(rule, _)| {
-                receive_markers.marker_index(rule.id).unwrap_or(usize::MAX)
+                receive_markers
+                    .same_update_marker_index(rule.id)
+                    .unwrap_or(usize::MAX)
             });
 
             self.ids_map.insert(archetype.id(), self.list.len());