+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Include receive markers in the protocol only when `MarkerConfig::apply_in_same_update` is enabled, allowing other markers to be registered only on clients.
+
 ## [0.43.0] - 2026-08-28
 
 ### Changed